requirements.txt
haystack-ai>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
transformers>=4.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
